Basic Connection/requirements.txt
langchain
langchain-community
langchain_core
langchain_openai
python-dotenv
pypdf
wikipedia
langchain-chroma
bs4
unstructured
faiss-cpu==1.8.0.post1

